fix: abandoned fractions in favor of decimal
The way fractions are working doesn't allow a lot of sequential operations on probability values to be consistent, because of value limit for numerator and denumerator. Thus, the simplest way to keep precision and stop caring about num. and denum. sizes is to throw them away all together and substitute fractions with simple decimal values. Works for hundreds of dice and seems to be fine for now.